fix ChemicalFilter copyWithAmount

+++ src/main/java/de/srendi/advancedperipherals/common/util/inventory/ChemicalFilter.java
@@ -135,7 +135,7 @@ public class ChemicalFilter extends GenericFilter<ChemicalStack> {
         return newFilter;
     }
 
-    public ChemicalFilter copyWithAmount(int amount) {
+    public ChemicalFilter copyWithAmount(long amount) {
         ChemicalFilter newFilter = this.copy();
         newFilter.amount = amount;
         return newFilter;
